+++ npm-dependency/index.html
@@ -22,7 +22,7 @@
 <head>
     <title>Open MCT as a Dependency Example</title>
     <script src="/node_modules/openmct/dist/openmct.js"></script>
-    <script src="/dist/openmct-example.js" defer></script>
+    <script src="openmct-example.js" defer></script>
 
     <link rel="icon" type="image/png" href="/node_modules/openmct/dist/favicons/favicon-96x96.png" sizes="96x96" type="image/x-icon">
     <link rel="icon" type="image/png" href="/node_modules/openmct/dist/favicons/favicon-32x32.png" sizes="32x32" type="image/x-icon">
