+++ npm-dependency/package.json
@@ -18,7 +18,7 @@
     "babel-loader": "9.1.0",
     "copy-webpack-plugin": "12.0.2",
     "css-loader": "6.10.0",
-    "mini-css-extract-plugin": "2.7.6",
+    "html-webpack-plugin": "5.6.0",
     "sass": "1.71.1",
     "sass-loader": "14.1.1",
     "source-map-loader": "4.0.1",
